Move room type converters into their own package
