+++ app/src/main/java/org/simple/clinic/medicalhistory/Answer.kt
@@ -4,7 +4,7 @@ import androidx.annotation.VisibleForTesting
 import androidx.room.TypeConverter
 import com.squareup.moshi.FromJson
 import com.squareup.moshi.ToJson
-import org.simple.clinic.util.SafeEnumTypeAdapter
+import org.simple.clinic.util.room.SafeEnumTypeAdapter
 
 sealed class Answer {
