+++ app/src/main/java/org/simple/clinic/overdue/Appointment.kt
@@ -12,7 +12,7 @@ import com.squareup.moshi.FromJson
 import com.squareup.moshi.ToJson
 import io.reactivex.Flowable
 import org.simple.clinic.patient.SyncStatus
-import org.simple.clinic.util.SafeEnumTypeAdapter
+import org.simple.clinic.util.room.SafeEnumTypeAdapter
 import org.threeten.bp.Instant
 import org.threeten.bp.LocalDate
 import java.util.UUID
