+++ app/src/main/java/org/simple/clinic/patient/PatientPhoneNumberType.kt
@@ -6,7 +6,7 @@ import androidx.room.TypeConverter
 import com.squareup.moshi.FromJson
 import com.squareup.moshi.ToJson
 import kotlinx.android.parcel.Parcelize
-import org.simple.clinic.util.SafeEnumTypeAdapter
+import org.simple.clinic.util.room.SafeEnumTypeAdapter
 
 sealed class PatientPhoneNumberType : Parcelable {
