+++ app/src/main/java/org/simple/clinic/patient/businessid/BusinessId.kt
@@ -19,7 +19,7 @@ import io.reactivex.Single
 import org.simple.clinic.patient.Patient
 import org.simple.clinic.patient.businessid.Identifier.IdentifierType
 import org.simple.clinic.patient.sync.BusinessIdPayload
-import org.simple.clinic.util.SafeEnumTypeAdapter
+import org.simple.clinic.util.room.SafeEnumTypeAdapter
 import org.threeten.bp.Instant
 import java.util.UUID
