+++ app/src/main/java/org/simple/clinic/patient/businessid/Identifier.kt
@@ -7,7 +7,7 @@ import androidx.room.TypeConverter
 import com.squareup.moshi.FromJson
 import com.squareup.moshi.ToJson
 import kotlinx.android.parcel.Parcelize
-import org.simple.clinic.util.SafeEnumTypeAdapter
+import org.simple.clinic.util.room.SafeEnumTypeAdapter
 
 @Parcelize
 data class Identifier(
