+++ app/src/main/java/org/simple/clinic/user/User.kt
@@ -9,7 +9,7 @@ import androidx.room.PrimaryKey
 import androidx.room.Query
 import androidx.room.Transaction
 import io.reactivex.Flowable
-import org.simple.clinic.util.RoomEnumTypeConverter
+import org.simple.clinic.util.room.RoomEnumTypeConverter
 import org.simple.clinic.util.UtcClock
 import org.threeten.bp.Instant
 import java.util.UUID
