+++ app/src/main/java/org/simple/clinic/util/room/LocalDateRoomTypeConverter.kt
@@ -1,4 +1,4 @@
-package org.simple.clinic.util
+package org.simple.clinic.util.room
 
 import androidx.room.TypeConverter
 import org.threeten.bp.LocalDate
