+++ app/src/main/java/org/simple/clinic/util/room/RoomEnumTypeConverter.kt
@@ -1,4 +1,4 @@
-package org.simple.clinic.util
+package org.simple.clinic.util.room
 
 import androidx.room.TypeConverter
